made boardController
